feat: split album logic into modules
